remove package pflag in pd-ctl
