+++ tools/pd-ctl/pdctl/ctl.go
@@ -14,8 +14,15 @@
 package pdctl
 
 import (
+	"fmt"
+	"github.com/pingcap/pd/v4/server"
+	"io"
 	"os"
+	"os/signal"
+	"syscall"
 
+	"github.com/chzyer/readline"
+	"github.com/mattn/go-shellwords"
 	"github.com/pingcap/pd/v4/tools/pd-ctl/pdctl/command"
 	"github.com/spf13/cobra"
 )
@@ -30,11 +37,43 @@ type CommandFlags struct {
 }
 
 var (
-	commandFlags = CommandFlags{}
+	commandFlags  = CommandFlags{}
+	Detach       bool
+	interact     bool
+	version      bool
 )
 
 func init() {
 	cobra.EnablePrefixMatching = true
+}
+
+func pdctlRun(cmd *cobra.Command, args []string) {
+	if version {
+		server.PrintPDInfo()
+		os.Exit(0)
+	}
+
+	sc := make(chan os.Signal, 1)
+	signal.Notify(sc,
+		syscall.SIGHUP,
+		syscall.SIGINT,
+		syscall.SIGTERM,
+		syscall.SIGQUIT)
+
+	go func() {
+		sig := <-sc
+		fmt.Printf("\nGot signal [%v] to exit.\n", sig)
+		switch sig {
+		case syscall.SIGTERM:
+			os.Exit(0)
+		default:
+			os.Exit(1)
+		}
+	}()
+
+	if interact {
+		loop()
+	}
 }
 
 // Start run Command
@@ -42,12 +81,18 @@ func Start(args []string) {
 	rootCmd := &cobra.Command{
 		Use:   "pd-ctl",
 		Short: "Placement Driver control",
+		Run:   pdctlRun,
 	}
 	rootCmd.PersistentFlags().StringVarP(&commandFlags.URL, "pd", "u", "http://127.0.0.1:2379", "pd address")
 	rootCmd.Flags().StringVar(&commandFlags.CAPath, "cacert", "", "path of file that contains list of trusted SSL CAs.")
 	rootCmd.Flags().StringVar(&commandFlags.CertPath, "cert", "", "path of file that contains X509 certificate in PEM format.")
 	rootCmd.Flags().StringVar(&commandFlags.KeyPath, "key", "", "path of file that contains X509 key in PEM format.")
 	rootCmd.PersistentFlags().BoolVarP(&commandFlags.Help, "help", "h", false, "Help message.")
+
+	rootCmd.Flags().BoolVarP(&Detach, "detach", "d", true, "Run pdctl without readline.")
+	rootCmd.Flags().BoolVarP(&interact, "interact", "i", false, "Run pdctl with readline.")
+	rootCmd.Flags().BoolVarP(&version, "version", "V", false, "Print version information and exit.")
+
 	rootCmd.AddCommand(
 		command.NewConfigCommand(),
 		command.NewRegionCommand(),
@@ -82,5 +127,44 @@ func Start(args []string) {
 
 	if err := rootCmd.Execute(); err != nil {
 		rootCmd.Println(err)
+	}
+}
+
+func loop() {
+	l, err := readline.NewEx(&readline.Config{
+		Prompt:            "\033[31m»\033[0m ",
+		HistoryFile:       "/tmp/readline.tmp",
+		InterruptPrompt:   "^C",
+		EOFPrompt:         "^D",
+		HistorySearchFold: true,
+	})
+	if err != nil {
+		panic(err)
+	}
+	defer l.Close()
+
+	for {
+		line, err := l.Readline()
+		if err != nil {
+			if err == readline.ErrInterrupt {
+				break
+			} else if err == io.EOF {
+				break
+			}
+			continue
+		}
+		if line == "exit" {
+			os.Exit(0)
+		}
+		args, err := shellwords.Parse(line)
+		if err != nil {
+			fmt.Printf("parse command err: %v\n", err)
+			continue
+		}
+		args = append(args, "-u", commandFlags.URL)
+		if commandFlags.CAPath != "" && commandFlags.CertPath != "" && commandFlags.KeyPath != "" {
+			args = append(args, "--cacert", commandFlags.CAPath, "--cert", commandFlags.CertPath, "--key", commandFlags.KeyPath)
+		}
+		Start(args)
 	}
 }
